Добавил возможность закрыть гитхабовские issue, чуть чуть отрефакторил код
